refactoring #3
