+++ packages/fuels-core/src/code_gen/mod.rs
@@ -5,3 +5,4 @@ pub mod bindings;
 pub mod custom_types_gen;
 pub mod docs_gen;
 pub mod functions_gen;
+mod resolved_type;
